+++ game/CTR/CTR_17_MatrixToRot.c
@@ -5,15 +5,15 @@
 /// @param rot - resulting rotation vector
 /// @param matrix - instance matrix
 /// @param flags - 3 bits boots + 2 bits some offset value (b0 - switch XZ, b1 - ?, b2 - negate the result)
-void CTR_MatrixToRot(SVECTOR *rot, MATRIX *matrix, u_int flags)
+void CTR_MatrixToRot(SVECTOR *rot, MATRIX *matrix, u32 flags)
 {
 	int iVar2;
 	int iVar3;
-	short *psVar5;
-	short *psVar11;
+	s16 *psVar5;
+	s16 *psVar11;
 
 	// bit 2 check - negate vector
-	u_int uVar7 = flags >> 2 & 1;
+	u32 uVar7 = flags >> 2 & 1;
 
 	// TODO: update global table def and use that instead of local
 	// TODO: figure out better names for whatever indices these are
@@ -28,11 +28,11 @@ void CTR_MatrixToRot(SVECTOR *rot, MATRIX *matrix, u_int flags)
 	char *table2 = (char *)0x8008d00C;
 
 	// take value from the first table
-	u_int t1value = (u_int)table1[flags >> 3 & 3];
+	u32 t1value = (u32)table1[flags >> 3 & 3];
 
 	// take two values from the second table
-	u_int t2value1 = (u_int)table2[t1value + uVar7];
-	u_int t2value2 = (u_int)table2[t1value - (uVar7 - 1)];
+	u32 t2value1 = (u32)table2[t1value + uVar7];
+	u32 t2value2 = (u32)table2[t1value - (uVar7 - 1)];
 
 	// bit 1 check - uses flipped matrix or smth?
 	if ((flags >> 1 & 1) == 1)
@@ -51,15 +51,15 @@ void CTR_MatrixToRot(SVECTOR *rot, MATRIX *matrix, u_int flags)
 		// based on sqrt result, calculate vector
 		if (iVar2 < 0x11)
 		{
-			rot->vx = (short)ratan2(-(int)matrix->m[t2value1][t2value2], (int)*(short *)((int)matrix->m + t2value1 * 8));
-			rot->vy = (short)ratan2(iVar2, (long)*(short *)((int)matrix->m + t1value * 8));
-			rot->vz = (short)0;
+			rot->vx = (s16)ratan2(-(int)matrix->m[t2value1][t2value2], (int)*(s16 *)((int)matrix->m + t2value1 * 8));
+			rot->vy = (s16)ratan2(iVar2, (s32) * (s16 *)((int)matrix->m + t1value * 8));
+			rot->vz = (s16)0;
 		}
 		else
 		{
-			rot->vx = (short)ratan2((int)*psVar5, (int)*psVar11);
-			rot->vy = (short)ratan2(iVar2, (int)*(short *)((int)matrix->m + t1value * 8));
-			rot->vz = (short)ratan2((int)matrix->m[t2value1][t1value], -(int)matrix->m[t2value2][t1value]);
+			rot->vx = (s16)ratan2((int)*psVar5, (int)*psVar11);
+			rot->vy = (s16)ratan2(iVar2, (int)*(s16 *)((int)matrix->m + t1value * 8));
+			rot->vz = (s16)ratan2((int)matrix->m[t2value1][t1value], -(int)matrix->m[t2value2][t1value]);
 		}
 	}
 	else
@@ -67,7 +67,7 @@ void CTR_MatrixToRot(SVECTOR *rot, MATRIX *matrix, u_int flags)
 		// TODO: there's definitely more elegant way for this
 		// probably it was smth like matrix->m[val1][val2]
 
-		psVar5 = (short *)((int)matrix->m + t1value * 8);
+		psVar5 = (s16 *)((int)matrix->m + t1value * 8);
 		iVar2 = (int)*psVar5;
 
 		psVar11 = matrix->m[t2value1] + t1value;
@@ -78,15 +78,15 @@ void CTR_MatrixToRot(SVECTOR *rot, MATRIX *matrix, u_int flags)
 		// based on sqrt result, calculate vector
 		if (iVar2 < 0x11)
 		{
-			rot->vx = (short)ratan2(-(int)matrix->m[t2value1][t2value2], (int)*(short *)((int)matrix->m + t2value1 * 8));
-			rot->vy = (short)ratan2(-(int)matrix->m[t2value2][t1value], iVar2);
-			rot->vz = (short)0;
+			rot->vx = (s16)ratan2(-(int)matrix->m[t2value1][t2value2], (int)*(s16 *)((int)matrix->m + t2value1 * 8));
+			rot->vy = (s16)ratan2(-(int)matrix->m[t2value2][t1value], iVar2);
+			rot->vz = (s16)0;
 		}
 		else
 		{
-			rot->vx = (short)ratan2((int)matrix->m[t2value2][t2value1], (int)*(short *)((int)matrix->m + t2value2 * 8));
-			rot->vy = (short)ratan2(-(int)matrix->m[t2value2][t1value], iVar2);
-			rot->vz = (short)ratan2((int)*psVar11, (int)*psVar5);
+			rot->vx = (s16)ratan2((int)matrix->m[t2value2][t2value1], (int)*(s16 *)((int)matrix->m + t2value2 * 8));
+			rot->vy = (s16)ratan2(-(int)matrix->m[t2value2][t1value], iVar2);
+			rot->vz = (s16)ratan2((int)*psVar11, (int)*psVar5);
 		}
 	}
 
@@ -109,5 +109,5 @@ void CTR_MatrixToRot(SVECTOR *rot, MATRIX *matrix, u_int flags)
 	}
 
 	// copy flags to pad
-	rot->pad = (short)flags;
+	rot->pad = (s16)flags;
 }
